requirements.txt
Django
django-cors-headers
djangorestframework
Faker
python-dateutil
pytz
six
sqlparse
text-unidecode
django-rest-auth
django-allauth
djangorestframework-simplejwt
django-import-export
pandas
